feat(seedvr2): tiled VAE encode/decode for memory-safe upscaling
Add tiled encode and decode to SeedVR2VAE, enabling 1024px+ output
without blowing up memory on intermediate tensors.

- tiledDecode: 64×64 latent tiles, stride 56, 64px cosine ramp blending
  in pixel space, CPU Float accumulator with weight normalization
- tiledEncode: 512×512 pixel tiles, stride 448, 8-unit cosine ramp
  blending in latent space
- Tile positions use while y <= dim - tileSize (matches Python mflux
  range(0, dim-tileSize+1, stride)) — avoids degenerate partial tiles
- Pipeline routes through tiled path; removes diagnostic prints from
  decoder; weight loader defaults to float16

Tested: 3B (28s, ~7GB) and 7B (~16GB) at 1024px produce output
matching the Python mflux reference.

+++ Sources/ZImage/Upscale/SeedVR2/Common/SeedVR2WeightLoader.swift
@@ -82,7 +82,7 @@ public enum SeedVR2WeightLoader {
     into transformer: SeedVR2Transformer,
     from directory: URL,
     fileName: String,
-    dtype: DType = .bfloat16,
+    dtype: DType = .float16,
     logger: Logger
   ) throws {
     let fileURL = directory.appendingPathComponent(fileName)
@@ -123,7 +123,7 @@ public enum SeedVR2WeightLoader {
     into vae: SeedVR2VAE,
     from directory: URL,
     fileName: String = vaeWeightFile,
-    dtype: DType = .bfloat16,
+    dtype: DType = .float16,
     logger: Logger
   ) throws {
     let fileURL = directory.appendingPathComponent(fileName)
